Add minimum battery level preference
This adds a preference for the minimum battery level the device will be
allowed to mine to when running on battery. If the device is mining on
battery and reaches this battery level, it will stop mining.

This also switches from hardcoded strings to resource strings for all
preference keys and resolves an issue with installing the widget.